requirements.txt
aiocontextvars==0.2.2
aiogram==2.7
aiohttp==3.6.2
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
Babel==2.8.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
gino==0.8.7
google-auth==1.13.1
gspread==3.4.2
idna==2.9
multidict==4.7.5
pip==19.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.13.0
pytz==2019.3
requests==2.23.0
rsa==4.0
setuptools==46.1.3
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.8
yarl==1.4.2
